Déplacement des vues dans le module py3o
